Correctly close subshell.
Blocks in a subshell must be closed before the subshell exits. This can
probably checked on exit from a subshell.  Traps set in the subshell
must be saved and restored then.  How about providing an `on_exit`
function as a generic endpoint for exiting a shell, i.e. just always
`trap "on_exit" EXIT` and define `on_exit` in the application.